+++ packages/core/src/components/EditableHeading/__stories__/EditableHeading.stories.tsx
@@ -54,6 +54,12 @@ export const Types = {
         />
         <EditableHeading
           type={EditableHeading.types.H1}
+          weight={EditableHeading.weights.MEDIUM}
+          value="H1 Medium"
+          className={styles.editableHeading}
+        />
+        <EditableHeading
+          type={EditableHeading.types.H1}
           weight={EditableHeading.weights.BOLD}
           value="H1 Bold"
           className={styles.editableHeading}
@@ -74,6 +80,12 @@ export const Types = {
         />
         <EditableHeading
           type={EditableHeading.types.H2}
+          weight={EditableHeading.weights.MEDIUM}
+          value="H2 Medium"
+          className={styles.editableHeading}
+        />
+        <EditableHeading
+          type={EditableHeading.types.H2}
           weight={EditableHeading.weights.BOLD}
           value="H2 Bold"
           className={styles.editableHeading}
@@ -90,6 +102,12 @@ export const Types = {
           type={EditableHeading.types.H3}
           weight={EditableHeading.weights.NORMAL}
           value="H3 Normal"
+          className={styles.editableHeading}
+        />
+        <EditableHeading
+          type={EditableHeading.types.H3}
+          weight={EditableHeading.weights.MEDIUM}
+          value="H3 Medium"
           className={styles.editableHeading}
         />
         <EditableHeading