+++ packages/core/src/components/Menu/MenuItem/MenuItem.tsx
@@ -11,6 +11,7 @@ import styles from "./MenuItem.module.scss";
 import BaseMenuItem from "./components/BaseMenuItem/BaseMenuItem";
 import MenuItemIcon from "./components/MenuItemIcon/MenuItemIcon";
 import { TooltipPositionsType } from "../../Tooltip/Tooltip.types";
+import { SubmenuPosition } from "./MenuItem.Types";
 
 export interface MenuItemProps extends VibeComponentProps {
   title?: string;
@@ -55,6 +56,7 @@ export interface MenuItemProps extends VibeComponentProps {
    */
   splitMenuItem?: boolean;
   "aria-label"?: AriaAttributes["aria-label"];
+  submenuPosition?: SubmenuPosition;
 }
 
 export interface MenuItemTitleComponentProps extends Omit<MenuItemProps, "title"> {