+++ packages/core/src/components/Menu/MenuItem/components/BaseMenuItem/BaseMenuItem.tsx
@@ -38,7 +38,8 @@ const BaseMenuItem = forwardRef(
       shouldScrollMenu,
       "data-testid": dataTestId,
       splitMenuItem = false,
-      children
+      children,
+      submenuPosition = "right"
     }: BaseMenuItemProps,
     ref: React.ForwardedRef<HTMLElement>
   ) => {
@@ -151,6 +152,7 @@ const BaseMenuItem = forwardRef(
               open={shouldShowSubMenu}
               onClose={closeSubMenu}
               autoFocusOnMount={!useDocumentEventListeners}
+              submenuPosition={submenuPosition}
             >
               {subMenu}
             </MenuItemSubMenu>