+++ packages/core/src/components/Table/Table/__stories__/Table.stories.helpers.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import Avatar from "../../../Avatar/Avatar";
 import { Calendar, Doc, Status } from "../../../Icon/Icons";
 import { LabelColor } from "../../../Label/LabelConstants";
-import { TableVirtualizedRow } from "../../TableVirtualizedBody/TableVirtualizedBody";
 import { ITableColumn } from "../Table";
 
 export const doAndDontIconsRuleColumns = [
@@ -118,7 +117,7 @@ export const emailTableData = [
   }
 ];
 
-export const emailColumns = [
+export const emailColumns: ITableColumn[] = [
   {
     id: "sentOn",
     title: "Sent on",
@@ -258,14 +257,14 @@ export const highlightableTableData = [
   }
 ];
 
-export const priorityColumnToLabelColor = {
+export const priorityColumnToLabelColor: { [key: string]: LabelColor } = {
   Urgent: LabelColor.NEGATIVE,
   High: LabelColor.DARK,
   Normal: LabelColor.PRIMARY,
   Low: LabelColor.POSITIVE
 };
 
-export const statusColumnToLabelColor = {
+export const statusColumnToLabelColor: { [key: string]: LabelColor } = {
   Sent: LabelColor.POSITIVE,
   Queued: LabelColor.DARK,
   Failed: LabelColor.NEGATIVE,
@@ -307,7 +306,7 @@ export const scrollTableColumns = [
   }
 ];
 
-export const virtualizedScrollTableData: TableVirtualizedRow[] = [...new Array(5000)].map((_, index) => ({
+export const virtualizedScrollTableData = [...new Array(5000)].map((_, index) => ({
   id: index.toString(),
   num: index,
   text: `This is line number ${index}`
@@ -325,7 +324,7 @@ export const virtualizedScrollTableColumns: ITableColumn[] = [
   }
 ];
 
-export function sort(columnId: string, sortState: "asc" | "desc" | "none", tableData: []) {
+export function sort<T>(columnId: keyof T, sortState: "asc" | "desc" | "none", tableData: T[]) {
   if (sortState === "asc") {
     return [...tableData].sort((a, b) => {
       return b[columnId] > a[columnId] ? 1 : -1;